mcp/server/requirements.txt
fastapi
uvicorn[standard]
requests
fastapi-mcp
line-bot-sdk==2.3.0